requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
appnope==0.1.3
    # via ipython
asgiref==3.5.2
    # via django
asttokens==2.0.5
    # via stack-data
attrs==21.4.0
    # via pytest
backcall==0.2.0
    # via ipython
black==22.6.0
    # via -r requirements.in
certifi==2022.6.15
    # via requests
charset-normalizer==2.1.0
    # via requests
click==8.1.3
    # via black
coverage[toml]==6.4.1
    # via pytest-cov
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
django==4.0.5
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.13.1
    # via -r requirements.in
executing==0.8.3
    # via stack-data
flake8==4.0.1
    # via -r requirements.in
idna==3.3
    # via requests
iniconfig==1.1.1
    # via pytest
ipdb==0.13.9
    # via -r requirements.in
ipython==8.4.0
    # via ipdb
jedi==0.18.1
    # via ipython
matplotlib-inline==0.1.3
    # via ipython
mccabe==0.6.1
    # via flake8
mypy==0.961
    # via -r requirements.in
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
packaging==21.3
    # via pytest
parso==0.8.3
    # via jedi
pathspec==0.9.0
    # via black
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
platformdirs==2.5.2
    # via black
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.30
    # via ipython
psycopg2==2.9.3
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
py==1.11.0
    # via pytest
pycodestyle==2.8.0
    # via flake8
pyflakes==2.4.0
    # via flake8
pygments==2.12.0
    # via ipython
pyparsing==3.0.9
    # via packaging
pytest==7.1.2
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-django
pytest-cov==3.0.0
    # via -r requirements.in
pytest-django==4.5.2
    # via -r requirements.in
pytz==2022.1
    # via djangorestframework
requests==2.28.1
    # via
    #   -r requirements.in
    #   requests-mock
requests-mock==1.9.3
    # via -r requirements.in
six==1.16.0
    # via requests-mock
sqlparse==0.4.2
    # via django
stack-data==0.3.0
    # via ipython
toml==0.10.2
    # via ipdb
tomli==2.0.1
    # via
    #   black
    #   coverage
    #   mypy
    #   pytest
traitlets==5.3.0
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.2.0
    # via
    #   black
    #   mypy
urllib3==1.26.9
    # via requests
wcwidth==0.2.5
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
